requirements.txt
python-daemon==2.2.4
requests==2.22.0
